+++ .storybook/main.ts
@@ -2,6 +2,10 @@ import type { StorybookConfig } from '@storybook/react-vite';
 
 const config: StorybookConfig = {
   stories: ['../src/**/*.mdx', '../src/**/*.stories.@(js|jsx|mjs|ts|tsx)'],
+  staticDirs: [
+    { from: '../public', to: '/' },
+    { from: '../docs', to: '/assets' },
+  ],
   addons: ['@storybook/addon-docs'],
   framework: {
     name: '@storybook/react-vite',